Allow uiContentFor to be used inside other directives
The removal of a directive's target element is not an action that is supported by the core
Angular $compile. It happens to work in general (as most uses of uiContentFor can show) but
fails when attempted inside another directive - due to the iterative nature of the compilation.

The next best thing is to replace the element with an empty text node. This is the cleanest way
to remove it without actually breaking the compile iteration.