Update content-for to wait for an $apply to look for yielders
When using uiContenFor in a template with an ngController, to project
content in an uiYieldTo located inside a directive, a race condition
prevents the uiContentFor from working: it executes before the uiYieldTo
has had time to register with Capture.

A simple solve is to do it in 2 trys: one right when setContentFor is called,
the other one in an $applyAsync callback so it is delayed enough for the
directive to have been compiled.